# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
  github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.SuccessDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        example: ""
        type: string
      errorMsg:
        example: ""
        type: string
      requestId:
        example: ""
        type: string
    type: object
...
  internal_transport_http_channel.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
...
  internal_transport_http_settings.PatchItem:
    properties:
      clear:
        type: boolean
      key:
        type: string
      namespace:
        type: string
      value:
        type: string
    required:
    - key
    - namespace
    type: object
  internal_transport_http_settings.PatchSettingsRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_transport_http_settings.PatchItem'
        minItems: 1
        type: array
    required:
    - items
    type: object
  internal_transport_http_usersettings.PatchSettingsRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
  internal_transport_http_usersettings.UserSettingsResponse:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    type: object
  internal_transport_http_usersettings.UserSettingsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponse'
      errorMsg:
        type: string
    type: object
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: 0.1.0
paths:
...
  /admin/llm/settings:
    get:
      consumes:
      - application/json
      description: superadmin 查询 LLM 全局设置列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.SuccessDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_channel.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员查询全局设置
      tags:
      - llm
  /admin/llm/settings/{key}:
    patch:
      consumes:
      - application/json
      description: superadmin 更新指定 LLM 全局设置项
      parameters:
      - description: 设置键
        in: path
        name: key
        required: true
        type: string
      - description: 设置值 {\
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.SuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_channel.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_channel.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_channel.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 管理员更新全局设置
      tags:
      - llm
...
  /admin/settings:
    get:
      description: 按 namespace 分组返回全部动态配置项
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询全部动态配置
      tags:
      - admin/settings
    patch:
      consumes:
      - application/json
      description: 批量更新动态配置并清除缓存，下次读取自动刷新
      parameters:
      - description: 更新项
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_settings.PatchSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 批量更新配置项
      tags:
      - admin/settings
  /admin/settings/{namespace}:
    get:
      description: 查询指定 namespace 下的全部配置项
      parameters:
      - description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询指定 namespace 的配置
      tags:
      - admin/settings
  /admin/settings/docling/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 Docling 运行状态
      tags:
      - admin/settings
...
  /admin/settings/mineru/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 MinerU 运行状态
      tags:
      - admin/settings
  /admin/settings/rapidocr/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 RapidOCR 运行状态
      tags:
      - admin/settings
  /admin/settings/rapidocr/runtime/restart:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 重启托管 RapidOCR
      tags:
      - admin/settings
  /admin/settings/rapidocr/runtime/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 启动托管 RapidOCR
      tags:
      - admin/settings
  /admin/settings/rapidocr/runtime/stop:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 停止托管 RapidOCR
      tags:
      - admin/settings
  /admin/settings/tesseract/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 Tesseract OCR 运行状态
      tags:
      - admin/settings
  /admin/settings/tika/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 Tika 运行状态
      tags:
      - admin/settings
  /admin/settings/tika/runtime/restart:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 重启托管 Tika
      tags:
      - admin/settings
  /admin/settings/tika/runtime/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 启动托管 Tika
      tags:
      - admin/settings
  /admin/settings/tika/runtime/stop:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 停止托管 Tika
      tags:
      - admin/settings
...
  /settings/model-option-policy:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询模型 options 透传策略
      tags:
      - settings
...
  /user/settings:
    get:
      description: 返回当前用户全部个人偏好配置，缺失项以默认值填充
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 获取当前用户的配置
      tags:
      - user/settings
    patch:
      consumes:
      - application/json
      description: 批量更新用户个人偏好配置，返回更新后的全量配置
      parameters:
      - description: 更新项
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_usersettings.PatchSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 更新当前用户的配置
      tags:
      - user/settings
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"